Extract stach watcher to utils
