+++ xawscf/commands/watch_stack.py
@@ -1,17 +1,15 @@
 '''
-Creates cloudformation stack
+Watches stack update events.
 '''
 
 from ..utils.client import get_client
 from ..utils.logger import log
 from ..utils.cloudformation import watch_stack_status
-from ..pattern.root import Root
 from ..pattern.pattern import get_pattern
 
 def run():
-    log('Creating stack')
+    log('Watching stack')
     pattern = get_pattern()
     stack_name = pattern.get('project')
     cf = get_client(pattern, 'cloudformation')
-    cf.create_stack(StackName=stack_name, TemplateBody=Root.TEMPLATE)
-    log('stack is created')
+    watch_stack_status(cf, stack_name)
